backend/requirements.txt
Django==3.2.12
djangorestframework==3.12.4
requests==2.28.2
django-cors-headers==3.14.0